+++ src/runez/__init__.py
@@ -2,7 +2,7 @@
 Friendly misc/utils/convenience library
 """
 
-from runez import click, config, file, program, serialize, system
+from runez import click, config, date, file, program, serialize, system
 from runez.colors import ActivateColors, ColorManager as color, uncolored
 from runez.colors.named import black, blue, brown, gray, green, orange, plain, purple, red, teal, white, yellow
 from runez.colors.named import blink, bold, dim, invert, italic, strikethrough, underline
@@ -12,7 +12,6 @@ from runez.convert import plural, represented_bytesize, represented_with_units
 from runez.convert import to_boolean, to_bytesize, to_float, to_int, unitized
 from runez.date import date_from_epoch, datetime_from_epoch, elapsed, local_timezone, represented_duration, \
     timezone, timezone_from_text, to_date, to_datetime, to_epoch, to_epoch_ms, to_seconds
-from runez.date import SECONDS_IN_ONE_DAY, SECONDS_IN_ONE_HOUR, SECONDS_IN_ONE_MINUTE, SECONDS_IN_ONE_YEAR, UTC
 from runez.file import copy, delete, move, readlines, symlink, TempFolder, touch, write
 from runez.logsetup import LogManager as log
 from runez.path import basename, ensure_folder, parent_folder
@@ -24,7 +23,7 @@ from runez.system import first_line, get_version, is_tty, PY2, resolved_path, WI
 
 __all__ = [
     "DRYRUN",
-    "click", "config", "file", "program", "serialize", "system",
+    "click", "config", "date", "file", "program", "serialize", "system",
     "ActivateColors", "color", "uncolored",
     "black", "blue", "brown", "gray", "green", "orange", "plain", "purple", "red", "teal", "white", "yellow",
     "blink", "bold", "dim", "invert", "italic", "strikethrough", "underline",
@@ -34,7 +33,6 @@ __all__ = [
     "to_boolean", "to_bytesize", "to_float", "to_int", "unitized",
     "date_from_epoch", "datetime_from_epoch", "elapsed", "local_timezone", "represented_duration",
     "timezone", "timezone_from_text", "to_date", "to_datetime", "to_epoch", "to_epoch_ms", "to_seconds",
-    "SECONDS_IN_ONE_DAY", "SECONDS_IN_ONE_HOUR", "SECONDS_IN_ONE_MINUTE", "SECONDS_IN_ONE_YEAR", "UTC",
     "copy", "delete", "move", "readlines", "symlink", "TempFolder", "touch", "write",
     "log",
     "basename", "ensure_folder", "parent_folder",